refactor: simplify SPOVotingChart component and integrate VotingChart
- Removed tooltip and hover state logic from SPOVotingChart for cleaner implementation.
- Integrated VotingChart component to handle voting data visualization.
- Updated DRepActivity page to include loading states and conditionally render DRepVotingChart.
- Adjusted SPOTable CSS for improved column width consistency.
