+++ apps/cardano-dashboard/src/pages/drep-activity.tsx
@@ -1,6 +1,10 @@
 import Head from "next/head";
+import { useDataContext } from "@/contexts/DataContext";
+import DRepVotingChart from "@/components/DRepVotingChart";
 
 export default function DRepActivity() {
+    const { governanceProposals, loading } = useDataContext();
+
     return (
         <div>
             <Head>
@@ -8,7 +12,13 @@ export default function DRepActivity() {
             </Head>
             <main>
                 <h1>DRep Activity</h1>
-                <p>Coming soon</p>
+                {loading.governanceProposals ? (
+                    <p>Loading DRep voting data...</p>
+                ) : governanceProposals && governanceProposals.length > 0 ? (
+                    <DRepVotingChart proposals={governanceProposals} />
+                ) : (
+                    <p>No DRep voting data available</p>
+                )}
             </main>
         </div>
     );
