--- apps/cardano-dashboard/src/styles/SPOVotingChart.module.css
@@ -1,131 +1,0 @@
-.chartContainer {
-    border-radius: 12px;
-    padding: 1.5rem;
-    backdrop-filter: blur(10px);
-    margin-bottom: 2rem;
-    position: relative;
-    width: 100%;
-}
-
-.chartContainer h2 {
-    color: var(--text-primary);
-    font-size: 1.25rem;
-    font-weight: 600;
-    margin: 0 0 1.5rem 0;
-}
-
-.chart {
-    display: flex;
-    align-items: flex-end;
-    gap: 2px;
-    height: 400px;
-    padding: 1rem 0;
-    margin-top: 3rem;
-    width: 100%;
-    justify-content: space-between;
-}
-
-.proposalBar {
-    flex: 1;
-    height: 100%;
-    display: flex;
-    align-items: flex-end;
-    min-width: 4px;
-    max-width: 20px;
-}
-
-.barContainer {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column-reverse;
-    border-radius: 4px;
-    overflow: hidden;
-    background: var(--background-secondary);
-}
-
-.barSegment {
-    width: 100%;
-    transition: opacity 0.2s ease;
-    cursor: pointer;
-    min-height: 4px;
-}
-
-.barSegment:hover {
-    opacity: 0.8;
-}
-
-.barSegment.yes {
-    background: var(--gradient-success);
-}
-
-.barSegment.no {
-    background: var(--gradient-danger);
-}
-
-.barSegment.abstain {
-    background: var(--gradient-neutral);
-}
-
-.tooltip {
-    position: absolute;
-    background: var(--gradient-info-action-4);
-    border-radius: 8px;
-    padding: 1rem;
-    color: var(--text-primary);
-    font-size: 0.875rem;
-    pointer-events: none;
-    z-index: 1000;
-    width: 360px;
-    backdrop-filter: blur(10px);
-    border: 1px solid var(--border-color);
-}
-
-.tooltipTitle {
-    font-weight: 600;
-    margin-bottom: 0.75rem;
-    color: var(--text-primary);
-    word-wrap: break-word;
-    white-space: normal;
-    line-height: 1.4;
-}
-
-.tooltipContent {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-}
-
-.tooltipRow {
-    display: grid;
-    grid-template-columns: 60px 1fr auto;
-    align-items: center;
-    gap: 1rem;
-}
-
-.tooltipRow span:first-child {
-    font-weight: 500;
-}
-
-.tooltipRow span:nth-child(2) {
-    color: var(--text-tertiary);
-    text-align: right;
-}
-
-.tooltipRow span:last-child {
-    color: var(--gradient-info-action-1);
-    min-width: 100px;
-    text-align: right;
-}
-
-.tooltipRow .yes {
-    color: var(--gradient-info-action-1);
-}
-
-.tooltipRow .no {
-    color: var(--gradient-parameter-change-1);
-}
-
-.tooltipRow .abstain {
-    color: var(--gradient-hard-fork-1);
-}
