+++ apps/cardano-dashboard/src/styles/VotingChart.module.css
@@ -1,0 +1,100 @@
+.chartContainer {
+    position: relative;
+    width: 100%;
+    padding: 20px;
+    background: var(--background-secondary);
+    border-radius: 8px;
+    box-shadow: 0 2px 4px var(--hover-shadow);
+}
+
+.chart {
+    display: flex;
+    align-items: flex-end;
+    gap: 4px;
+    height: 300px;
+    margin-top: 20px;
+}
+
+.proposalBar {
+    flex: 1;
+    height: 100%;
+    min-width: 20px;
+    cursor: pointer;
+    transition: transform 0.2s ease;
+}
+
+.proposalBar:hover {
+    transform: scale(1.05);
+}
+
+.barContainer {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    gap: 1px;
+}
+
+.barSegment {
+    width: 100%;
+    transition: height 0.3s ease;
+}
+
+.yes {
+    background: var(--gradient-success);
+}
+
+.no {
+    background: var(--gradient-danger);
+}
+
+.abstain {
+    background: var(--gradient-neutral);
+}
+
+.tooltip {
+    position: absolute;
+    background: var(--background-primary);
+    border: 1px solid var(--border-color);
+    border-radius: 4px;
+    padding: 12px;
+    min-width: 300px;
+    box-shadow: 0 4px 6px var(--hover-shadow);
+    z-index: 1000;
+    pointer-events: none;
+}
+
+.tooltipTitle {
+    font-weight: 600;
+    margin-bottom: 8px;
+    color: var(--text-primary);
+}
+
+.tooltipContent {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.tooltipRow {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: 12px;
+}
+
+.tooltipRow span:first-child {
+    min-width: 60px;
+    font-weight: 500;
+}
+
+.tooltipRow span:nth-child(2) {
+    flex: 1;
+    text-align: right;
+}
+
+.tooltipRow span:last-child {
+    min-width: 100px;
+    text-align: right;
+    font-family: monospace;
+}
